fix: remove all wrong points
